Added Modes for every time interval possible
Satellite now goes through modes as it travels around the Earth. The satellite goes to Imaging when it is taking pictures, Data Processing for 30 minutes after it takes pictures, Data Downlink when above the UGA ground station, and Cruise mode when it is not in any other mode and just charging the solar panels.